+++ src/widget/time/timepicker-extended.js
@@ -32,7 +32,7 @@ TimepickerExtended.prototype._init = function() {
     var $timeI = $( this.element );
     var timeVal = $( this.element ).val();
     var $fakeTime = $( '<div class="widget timepicker">' +
-        '<input class="ignore timepicker-default input-small" type="text" value="' + timeVal + '" placeholder="hh:mm" />' +
+        '<input class="ignore timepicker-default" type="text" value="' + timeVal + '" placeholder="hh:mm" />' +
         '<button class="btn-icon-only btn-reset" type="button"><i class="icon icon-refresh"> </i></button></div>' );
     var $resetBtn = $fakeTime.find( '.btn-reset' );
     var $fakeTimeI = $fakeTime.find( 'input' );
@@ -40,24 +40,19 @@ TimepickerExtended.prototype._init = function() {
     $timeI.hide().after( $fakeTime );
 
     $fakeTimeI.timepicker( {
-            defaultTime: ( timeVal.length > 0 ) ? timeVal : 'current',
+            defaultTime: ( timeVal.length > 0 ) ? timeVal : false,
             showMeridian: false
         } ).val( timeVal )
         //the time picker itself has input elements
         .closest( '.widget' ).find( 'input' ).addClass( 'ignore' );
 
     $fakeTimeI.on( 'change', function() {
-        var val = this.value;
-        // the following line can be removed if https://github.com/jdewit/bootstrap-timepicker/issues/202 gets approved
-        val = /^[0-9]:/.test( val ) ? '0' + val : val;
-        // add 00 minutes if they are missing (probably a bug in bootstrap timepicker)
-        val = /^[0-9]{2}:$/.test( val ) ? val + '00' : val;
-        $timeI.val( val ).trigger( 'change' ).blur();
+        $timeI.val( this.value ).trigger( 'change' );
         return false;
     } );
 
     // reset button
-    $fakeTimeReset.on( 'click', function() {
+    $resetBtn.on( 'click', function() {
         $fakeTimeI.val( '' ).trigger( 'change' );
     } );